[level 0] Title: 그림 확대, Time: 0.07 ms, Memory: 4.7 MB -BaekjoonHub

+++ 프로그래머스/0/181836. 그림 확대/그림 확대.cpp
@@ -7,8 +7,8 @@ vector<string> solution(vector<string> picture, int k) {
     vector<string> answer;
     string temp;
     
-    for(int i{}; i < picture.size(); i++)   {
-        for(int j{}; j < picture[i].size(); j++)    {
+    for(size_t i{}; i < picture.size(); i++)   {
+        for(size_t j{}; j < picture[i].size(); j++)    {
             for(int l{}; l < k; l++)    {
                 temp += picture[i][j];
             }
